fearless_simd::generated::simd_types

Struct mask16x16

#[repr(C, align(32))]
pub struct mask16x16<S: Simd> {
    pub(crate) val: S::mask16x16,
    pub simd: S,
}

A SIMD mask of 16 16-bit elements.

When created from a comparison operation, and as it should be used in a Self::select operation, each element will be all ones if it’s “true”, and all zeroes if it’s “false”.

Fields
val: S::mask16x16simd: S