fearless_simd

Struct mask32x4

#[repr(C, align(16))]
pub struct mask32x4<S: Simd> {
    pub(crate) val: S::mask32x4,
    pub simd: S,
}

A SIMD mask of 4 32-bit elements.

When created from a comparison operation, and as it should be used in a Self::select operation, each element will be all ones if it’s “true”, and all zeroes if it’s “false”.

Fields
val: S::mask32x4simd: S