fearless_simd

Struct mask8x64

#[repr(C, align(64))]
pub struct mask8x64<S: Simd> {
    pub(crate) val: S::mask8x64,
    pub simd: S,
}

A SIMD mask of 64 8-bit elements.

When created from a comparison operation, and as it should be used in a Self::select operation, each element will be all ones if it’s “true”, and all zeroes if it’s “false”.

Fields
val: S::mask8x64simd: S