fearless_simd

Module support

Structs

Aligned128

Wrapper for internal native vector types that gives them 128-bit alignment.

Aligned256

Wrapper for internal native vector types that gives them 256-bit alignment.

Aligned512

Wrapper for internal native vector types that gives them 512-bit alignment.

Functions

cross_block_slide_blocks_at 🔒

Selects the input operands to be used for slignr/vext/etc. when computing a single output block for cross-block “slide” operations. Extracts from [a : b].

simd_debug_impl 🔒

The actual Debug implementation for all SimdBase types. This only needs to be monomorphized once per element type, rather than once per vector type.