fearless_simd

Trait Select

pub trait Select<T> {
    // Required method
    fn select(self, if_true: T, if_false: T) -> T;
}

Element-wise selection between two SIMD vectors using self.

Required Methods

fn select(self, if_true: T, if_false: T) -> T

For each element of this mask, select the first operand if the element is all ones, and select the second operand if the element is all zeroes.

If a mask element is not all ones or all zeroes, the result is unspecified. It may vary depending on architecture, feature level, the mask elements’ width, the mask vector’s width, or library version.